requirements.txt
notebook
PyPDF2
python-docx
python-pptx
nltk
torch
tf-keras
transformers
scikit-learn
tqdm